# internal/api/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Motus API
  version: "0.1.0"
  description: GPS tracking and fleet management API
servers:
  - url: /
    description: This server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie set after login
    bearerAuth:
      type: http
      scheme: bearer
      description: API key token
    xAuthToken:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: iOS PWA fallback — pass session_id value here
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "resource not found"
...
    Geofence:
      type: object
      required: [id, name, area, attributes, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        description: { type: string }
        area: { type: string, description: "WKT geometry (Traccar-compatible)" }
        geometry: { type: string, description: "GeoJSON geometry" }
        calendarId: { type: integer, format: int64, nullable: true }
        ownerName: { type: string }
        attributes:
          type: object
          additionalProperties: {}
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    GeofenceInput:
      type: object
      required: [name, area]
      properties:
        name: { type: string }
        description: { type: string }
        area: { type: string }
        geometry: { type: string }
        calendarId: { type: integer, format: int64 }
        attributes:
          type: object
          additionalProperties: {}
...
paths:
...
  /api/geofences:
    get:
      operationId: listGeofences
      summary: List geofences for the authenticated user
      tags: [geofences]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      responses:
        "200":
          description: Geofence list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Geofence'
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createGeofence
      summary: Create a geofence
      tags: [geofences]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeofenceInput'
      responses:
        "201":
          description: Geofence created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /api/geofences/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      operationId: getGeofence
      summary: Get a geofence by ID
      tags: [geofences]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      responses:
        "200":
          description: Geofence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateGeofence
      summary: Update a geofence
      tags: [geofences]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeofenceInput'
      responses:
        "200":
          description: Updated geofence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteGeofence
      summary: Delete a geofence
      tags: [geofences]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      responses:
        "204":
          description: Deleted
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
...
  /api/admin/geofences:
    get:
      operationId: adminListGeofences
      summary: Admin — list all geofences across all users
      tags: [admin]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      responses:
        "200":
          description: All geofences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Geofence'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
...